requirements.txt
cassandra-driver
openai
dynaconf
loguru